requirements.txt
graphviz
google-generativeai
PyPDF2
requests
pandas
matplotlib
seaborn
plotly
